+++ CHANGELOG.md
@@ -37,6 +37,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added CMReportingServicePoint Resource
 - Added CMSystemDiscovery Resource
 - Added CMPxeDistributionPoint Resource
+- Added CMPullDistributionPoint Resource
+- Added ConvertTo-AnyCimInstance to the ResourceHelper
 
 ### Changed